refactor: muda nome de caso de uso que cria um novo saque
