+++ app/Application/CreateWithdrawal/WithdrawalCreator.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Application\WithdrawFunds;
+namespace App\Application\CreateWithdrawal;
 
 use App\Domain\Entity\Withdrawal;
 use App\Domain\Entity\WithdrawalMethod;
@@ -9,7 +9,7 @@ use App\Domain\ValueObject\Withdrawal\WithdrawalSchedule;
 use App\Repository\AccountRepository;
 use Throwable;
 
-class FundsWithdrawer
+class WithdrawalCreator
 {
     public function __construct(
         private AccountRepository $accountRepository,
